Export the commonjs module for the web
