+++ lib/utils.js
@@ -1,16 +1,3 @@
-// Load window.performance in the browser, perf_hooks in node, and fall back on Date
-const requirePerfHooks = () => {
-    try {
-        const result = require('perf_hooks');
-        if (result && result.performance) {
-            return result;
-        }
-    } catch (e) {
-        
-    }
-};
-const { performance = Date } = requirePerfHooks() || window || {};
-
 const noop = () => {};
 const ifDefined = (cb = noop) => (input) => {
     if (input !== undefined) {
@@ -75,7 +62,5 @@ module.exports = {
     binaryStringToArray,
     arrayToBinaryString,
     getHashedPassword,
-    stripArmor,
-    updateServerTime,
-    serverTime
+    stripArmor
 };
